# config/api.yml
swagger: '2.0'
info:
  title: Gammabet API
  description: >-
    This API allows complete interaction with Gammabet, built with
    mobile clients in mind.
  contact:
    email: aksingh@grepruby.com
  version: '1.0'
host: 'localhost:3000'
basePath: "/api/v1/"
...
schemes:
- "http"
paths:
...
  /matches:
    get:
      tags:
      - "match"
      summary: "Return all matches"
      description: "Return all matches"
      operationId: "matchescontroller"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
        - name: 'scope'
          in: 'query'
          description: 'Scope of matches [live, upcoming, last_minutes, highlight]'
          required: false
          type: 'string'
          enum: ['live', 'upcoming', 'last_minutes', 'highlight']
          default: 'all'
          allowEmptyValue: false
        - name: 'page'
          in: 'query'
          description: 'The paginated list page.'
          required: false
          type: 'integer'
          allowEmptyValue: false
        - name: 'per_page'
          in: 'query'
          description: 'The paginated list limit.'
          required: false
          type: 'integer'
          allowEmptyValue: false
        - name: 'country'
          in: 'query'
          description: 'country'
          required: false
          type: 'string'
          allowEmptyValue: false
        - name: 'continent'
          in: 'query'
          description: 'continent'
          required: false
          type: 'string'
          allowEmptyValue: false
        - name: 'tournament_id'
          in: 'query'
          description: 'tournament'
          required: false
          type: 'string'
          allowEmptyValue: false
        - name: 'sport_id'
          in: 'query'
          description: 'sport'
          required: false
          type: 'string'
          allowEmptyValue: false
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Matches"
        401:
          description: 'Authorization information is missing or invalid.'
      security:
        - JWT: []
...
  /hold/confirm:
    post:
      tags:
      - 'bet'
      summary: 'Confirm a hold bet'
      description: 'Confirm a hold bet'
      operationId: 'HoldController'
      produces:
      - 'application/xml'
      - 'application/json'
      parameters:
        - name: "body"
          in: "body"
          description: "List of bets to confirm"
          required: true
          schema:
            type: 'object'
            properties:
              bets:
                type: 'array'
                title: 'bets'
                items:
                  properties:
                    id:
                      type: 'integer'
                      title: 'id'
      responses:
        200:
          description: "Bet is confirmed."
        401:
          description: 'Authorization information is missing or invalid.'
      security:
        - JWT: []
...
  /topics:
    get:
      tags:
      - "topic"
      summary: "Return all topics"
      description: "Return all topics"
      operationId: "topicscontroller"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Topic"
        401:
          description: 'Authorization information is missing or invalid.'
      security:
        - JWT: []
...
securityDefinitions:
  JWT:
    description: ""
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  Topic:
    type: 'object'
    properties:
      id:
        type: 'integer'
        format: 'int64'
        example: 1'
      name:
        type: 'string'
        example: 'fcb'
      enabled:
        type: 'boolean'
        example: true
...
  Sport:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
        example: "Soccer"
      favourite:
        type: 'boolean'
        example: true
      favourite_id:
        type: 'integer'
        example: 1
    xml:
      name: "Sport"
...
  Match:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        example: 1
      status:
        type: "string"
        example: "not_started"
      schedule_at:
        type: "string"
        format: "date-time"
        example: Wed, 07 Feb 2018 15:00:00 UTC +00:00
      venue:
        $ref: "#/definitions/Venue"
      teams:
        type: 'array'
        items:
          type: 'object'
          properties:
            id:
              type: 'integer'
              format: 'int64'
              example: 1
            name:
              type: 'string'
              example: 'Atletico Madrid'
            acronym:
              type: 'string'
              example: 'ALT'
      sport:
        $ref: "#/definitions/Sport"
      tournament:
        type: 'object'
        properties:
          id:
            type: 'integer'
            format: 'int64'
            example: 1
          name:
            type: 'string'
            example: 'Eredivisie'
...
  Venue:
    type: 'object'
    properties:
      id:
        type: 'integer'
        format: 'int64'
        example: 1
      name:
        type: 'string'
        example: 'Vitesse Arnhem'
      city_name:
        type: 'string'
        example: 'Den Haag'
      country_name:
        type: 'string'
        example: 'Netherlands'
...
  Matches:
    type: 'array'
    items:
      $ref: '#/definitions/Match'
...